fix: renamed StartupChecks -> system-checks
